backend/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
python-multipart
pillow
openai
transformers
torch
accelerate
easyocr